requirements.txt
# Core Dependencies
streamlit>=1.31.0
python-dotenv>=1.0.1

# PDF Processing
PyPDF2>=3.0.1

# Google Gemini AI
google-generativeai>=0.3.2

# LangChain
langchain>=0.1.6
langchain-community>=0.0.19
langchain-google-genai>=0.0.6

# Vector Store and Embeddings
faiss-cpu>=1.8.0
sentence-transformers>=2.3.1

# Additional Dependencies
huggingface-hub>=0.20.3



